+++ plugins/source/aws/resources/services/elbv2/listener_certificates.go
@@ -1,13 +1,17 @@
 package elbv2
 
 import (
+	"context"
+
+	"github.com/aws/aws-sdk-go-v2/aws"
+	elbv2 "github.com/aws/aws-sdk-go-v2/service/elasticloadbalancingv2"
 	"github.com/aws/aws-sdk-go-v2/service/elasticloadbalancingv2/types"
 	"github.com/cloudquery/cloudquery/plugins/source/aws/client"
 	"github.com/cloudquery/plugin-sdk/schema"
 	"github.com/cloudquery/plugin-sdk/transformers"
 )
 
-func ListenerCertificates() *schema.Table {
+func listenerCertificates() *schema.Table {
 	tableName := "aws_elbv2_listener_certificates"
 	return &schema.Table{
 		Name:        tableName,
@@ -25,4 +29,26 @@ func ListenerCertificates() *schema.Table {
 			},
 		},
 	}
+}
+
+func fetchElbv2ListenerCertificates(ctx context.Context, meta schema.ClientMeta, parent *schema.Resource, res chan<- any) error {
+	c := meta.(*client.Client)
+	region := c.Region
+	svc := c.Services().Elasticloadbalancingv2
+	listener := parent.Item.(types.Listener)
+	config := elbv2.DescribeListenerCertificatesInput{ListenerArn: listener.ListenerArn}
+	for {
+		response, err := svc.DescribeListenerCertificates(ctx, &config, func(options *elbv2.Options) {
+			options.Region = region
+		})
+		if err != nil {
+			return err
+		}
+		res <- response.Certificates
+		if aws.ToString(response.NextMarker) == "" {
+			break
+		}
+		config.Marker = response.NextMarker
+	}
+	return nil
 }
