+++ plugins/source/aws/resources/services/elbv2/listeners.go
@@ -1,13 +1,17 @@
 package elbv2
 
 import (
+	"context"
+
+	"github.com/aws/aws-sdk-go-v2/aws"
+	elbv2 "github.com/aws/aws-sdk-go-v2/service/elasticloadbalancingv2"
 	"github.com/aws/aws-sdk-go-v2/service/elasticloadbalancingv2/types"
 	"github.com/cloudquery/cloudquery/plugins/source/aws/client"
 	"github.com/cloudquery/plugin-sdk/schema"
 	"github.com/cloudquery/plugin-sdk/transformers"
 )
 
-func Listeners() *schema.Table {
+func listeners() *schema.Table {
 	tableName := "aws_elbv2_listeners"
 	return &schema.Table{
 		Name:        tableName,
@@ -34,7 +38,58 @@ func Listeners() *schema.Table {
 		},
 
 		Relations: []*schema.Table{
-			ListenerCertificates(),
+			listenerCertificates(),
 		},
 	}
+}
+
+func fetchElbv2Listeners(ctx context.Context, meta schema.ClientMeta, parent *schema.Resource, res chan<- any) error {
+	lb := parent.Item.(types.LoadBalancer)
+	config := elbv2.DescribeListenersInput{
+		LoadBalancerArn: lb.LoadBalancerArn,
+	}
+	c := meta.(*client.Client)
+	svc := c.Services().Elasticloadbalancingv2
+	for {
+		response, err := svc.DescribeListeners(ctx, &config)
+		if err != nil {
+			if c.IsNotFoundError(err) {
+				return nil
+			}
+			return err
+		}
+		res <- response.Listeners
+		if aws.ToString(response.NextMarker) == "" {
+			break
+		}
+		config.Marker = response.NextMarker
+	}
+	return nil
+}
+
+func resolveElbv2listenerTags(ctx context.Context, meta schema.ClientMeta, resource *schema.Resource, c schema.Column) error {
+	region := meta.(*client.Client).Region
+	svc := meta.(*client.Client).Services().Elasticloadbalancingv2
+	listener := resource.Item.(types.Listener)
+	tagsOutput, err := svc.DescribeTags(ctx, &elbv2.DescribeTagsInput{
+		ResourceArns: []string{
+			*listener.ListenerArn,
+		},
+	}, func(o *elbv2.Options) {
+		o.Region = region
+	})
+	if err != nil {
+		return err
+	}
+	if len(tagsOutput.TagDescriptions) == 0 {
+		return nil
+	}
+	tags := make(map[string]*string)
+	for _, td := range tagsOutput.TagDescriptions {
+		for _, s := range td.Tags {
+			tags[*s.Key] = s.Value
+		}
+	}
+
+	return resource.Set(c.Name, tags)
 }
