+++ plugins/source/aws/resources/services/elbv2/load_balancer_attributes.go
@@ -1,13 +1,16 @@
 package elbv2
 
 import (
+	"context"
+
+	elbv2 "github.com/aws/aws-sdk-go-v2/service/elasticloadbalancingv2"
 	"github.com/aws/aws-sdk-go-v2/service/elasticloadbalancingv2/types"
 	"github.com/cloudquery/cloudquery/plugins/source/aws/client"
 	"github.com/cloudquery/plugin-sdk/schema"
 	"github.com/cloudquery/plugin-sdk/transformers"
 )
 
-func LoadBalancerAttributes() *schema.Table {
+func loadBalancerAttributes() *schema.Table {
 	tableName := "aws_elbv2_load_balancer_attributes"
 	return &schema.Table{
 		Name:        tableName,
@@ -25,4 +28,16 @@ func LoadBalancerAttributes() *schema.Table {
 			},
 		},
 	}
+}
+
+func fetchElbv2LoadBalancerAttributes(ctx context.Context, meta schema.ClientMeta, parent *schema.Resource, res chan<- any) error {
+	lb := parent.Item.(types.LoadBalancer)
+	c := meta.(*client.Client)
+	svc := c.Services().Elasticloadbalancingv2
+	result, err := svc.DescribeLoadBalancerAttributes(ctx, &elbv2.DescribeLoadBalancerAttributesInput{LoadBalancerArn: lb.LoadBalancerArn})
+	if err != nil {
+		return err
+	}
+	res <- result.Attributes
+	return nil
 }
