+++ plugins/source/aws/resources/services/elbv2/target_group_target_health_descriptions.go
@@ -1,13 +1,16 @@
 package elbv2
 
 import (
+	"context"
+
+	elbv2 "github.com/aws/aws-sdk-go-v2/service/elasticloadbalancingv2"
 	"github.com/aws/aws-sdk-go-v2/service/elasticloadbalancingv2/types"
 	"github.com/cloudquery/cloudquery/plugins/source/aws/client"
 	"github.com/cloudquery/plugin-sdk/schema"
 	"github.com/cloudquery/plugin-sdk/transformers"
 )
 
-func TargetGroupTargetHealthDescriptions() *schema.Table {
+func targetGroupTargetHealthDescriptions() *schema.Table {
 	tableName := "aws_elbv2_target_group_target_health_descriptions"
 	return &schema.Table{
 		Name:        tableName,
@@ -25,4 +28,21 @@ func TargetGroupTargetHealthDescriptions() *schema.Table {
 			},
 		},
 	}
+}
+
+func fetchElbv2TargetGroupTargetHealthDescriptions(ctx context.Context, meta schema.ClientMeta, parent *schema.Resource, res chan<- any) error {
+	cl := meta.(*client.Client)
+	svc := cl.Services().Elasticloadbalancingv2
+	tg := parent.Item.(types.TargetGroup)
+	response, err := svc.DescribeTargetHealth(ctx, &elbv2.DescribeTargetHealthInput{
+		TargetGroupArn: tg.TargetGroupArn,
+	})
+	if err != nil {
+		if cl.IsNotFoundError(err) {
+			return nil
+		}
+		return err
+	}
+	res <- response.TargetHealthDescriptions
+	return nil
 }
