--- plugins/source/aws/resources/services/elbv2/target_groups_fetch.go
@@ -1,73 +1,0 @@
-package elbv2
-
-import (
-	"context"
-
-	"github.com/aws/aws-sdk-go-v2/aws"
-	elbv2 "github.com/aws/aws-sdk-go-v2/service/elasticloadbalancingv2"
-	"github.com/aws/aws-sdk-go-v2/service/elasticloadbalancingv2/types"
-	"github.com/cloudquery/cloudquery/plugins/source/aws/client"
-	"github.com/cloudquery/plugin-sdk/schema"
-)
-
-func fetchElbv2TargetGroups(ctx context.Context, meta schema.ClientMeta, parent *schema.Resource, res chan<- any) error {
-	var config elbv2.DescribeTargetGroupsInput
-	c := meta.(*client.Client)
-	svc := c.Services().Elasticloadbalancingv2
-	for {
-		response, err := svc.DescribeTargetGroups(ctx, &config)
-		if err != nil {
-			return err
-		}
-		res <- response.TargetGroups
-		if aws.ToString(response.NextMarker) == "" {
-			break
-		}
-		config.Marker = response.NextMarker
-	}
-	return nil
-}
-func resolveElbv2targetGroupTags(ctx context.Context, meta schema.ClientMeta, resource *schema.Resource, c schema.Column) error {
-	cl := meta.(*client.Client)
-	region := meta.(*client.Client).Region
-	svc := meta.(*client.Client).Services().Elasticloadbalancingv2
-	targetGroup := resource.Item.(types.TargetGroup)
-	tagsOutput, err := svc.DescribeTags(ctx, &elbv2.DescribeTagsInput{
-		ResourceArns: []string{
-			*targetGroup.TargetGroupArn,
-		},
-	}, func(o *elbv2.Options) {
-		o.Region = region
-	})
-	if err != nil {
-		if cl.IsNotFoundError(err) {
-			return nil
-		}
-		return err
-	}
-	if len(tagsOutput.TagDescriptions) == 0 {
-		return nil
-	}
-	tags := make(map[string]*string)
-	for _, s := range tagsOutput.TagDescriptions[0].Tags {
-		tags[*s.Key] = s.Value
-	}
-	return resource.Set(c.Name, tags)
-}
-
-func fetchElbv2TargetGroupTargetHealthDescriptions(ctx context.Context, meta schema.ClientMeta, parent *schema.Resource, res chan<- any) error {
-	cl := meta.(*client.Client)
-	svc := cl.Services().Elasticloadbalancingv2
-	tg := parent.Item.(types.TargetGroup)
-	response, err := svc.DescribeTargetHealth(ctx, &elbv2.DescribeTargetHealthInput{
-		TargetGroupArn: tg.TargetGroupArn,
-	})
-	if err != nil {
-		if cl.IsNotFoundError(err) {
-			return nil
-		}
-		return err
-	}
-	res <- response.TargetHealthDescriptions
-	return nil
-}
